backend/api-service/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.5
Brlapi==0.8.4
cachetools==5.2.0
certifi==2022.9.24
chardet==5.1.0
click==8.1.3
colorama==0.4.6
command-not-found==0.3
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.8.0
distro-info==1.5+ubuntu0.23.4.1
dnspython==2.3.0
duplicity==0.8.22
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.17.3
future==0.18.2
google-api-python-client==1.7.12
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==2.0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.4
httptools==0.6.1
httpx==0.27.0
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jose==1.0.0
keyring==23.9.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.24.0
Mako==1.2.4.dev0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
orjson==3.10.3
paramiko==2.12.0
pexpect==4.8.0
Pillow==9.4.0
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.3
pydantic-settings==2.3.0
pydantic_core==2.18.4
PyDrive==1.3.1
PyDrive2==0.0.0
Pygments==2.14.0
PyGObject==3.44.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-apt==2.5.3+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-gnupg==0.4.9
python-multipart==0.0.9
pythondialog==3.5.1
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.3.1
rsa==4.8
SecretStorage==3.3.3
setuptools-git==1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
speedtest-cli==2.1.3
SQLAlchemy==1.4.46
starlette==0.37.2
systemd-python==235
typer==0.12.3
typing_extensions==4.12.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.16
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
websockets==12.0
xdg==5
xkit==0.0.0
zipp==1.0.0
